[CARBONDATA-2990] Fixed JVM crash when rebuilding bloom datamap
Problem: while rebuilding the datamap it access the datamap store so it builds datamap and store in unsafe onheap storage. But while closing the
reader it frees all memory acquired during that task. Since acquired memory is onheap but releasing the memory with offheap allocator it crashes the jvm.

Solution: Maintain the type of memory acquired in the memory block itself and get the allocator as per the memory type and release it.

This closes #2793