[CARBONDATA-2994] unify badrecordpath property name for create and load
Problem:
Currently bad records path can be specified in create and load. In create the property name is bad_records_path and load is bad_record_path. This can cause confusion for the user.

Solution: Use bad_record_path as the property for create so that both load and create use the same name.

This closes #2799